New branch with UUID fs2v8bk1n41qrywk3dg2axkzv
Initial branch nickname: debrutestrap

Initial version

The goal of this project is to replace Debian's "debootstrap" which
never worked to my full satisfaction.

The idea is to start with nothing else than a copy of /var/apt/archives
and figure out what and how to install everything necessary for a bare
minimum system by oneself.

Such a system is defined by providing a working "dpkg" with all its
dependencies already installed manually somehow (which is the goal of
this project to find out).

Then one can re-install all those packages again, fixing things those
primitive scripts might have missed. And from then on, the system should
be self-sufficient to install more and more packages until everything is
up to the user's expectations.